+++ Dropdown/Dropdown.js
@@ -17,7 +17,7 @@
  */
 
 import EnactPropTypes from '@enact/core/internal/prop-types';
-import {handle, forKey, forward, forProp, not} from '@enact/core/handle';
+import {handle, forward, forProp, not} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import {extractAriaProps} from '@enact/core/util';
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
@@ -224,19 +224,6 @@ const DropdownBase = kind({
 	},
 
 	handlers: {
-		onKeyDown: handle(
-			forward('onKeyDown'),
-			(ev, props) => {
-				const {rtl} = props;
-				const isLeft = forKey('left', ev, props);
-				const isRight = forKey('right', ev, props);
-				const isLeftMovement = !rtl && isLeft || rtl && isRight;
-				const isRightMovement = !rtl && isRight || rtl && isLeft;
-
-				return isLeftMovement && typeof ev.target.dataset.index !== 'undefined' || isRightMovement;
-			},
-			forward('onClose')
-		),
 		onSelect: handle(
 			forward('onSelect'),
 			forward('onClose')
@@ -255,7 +242,7 @@ const DropdownBase = kind({
 	},
 
 	computed: {
-		'aria-labelledby': ({id, title}) => (title ? `${id}_title` : void 0),
+		ariaLabelledBy: ({id, title}) => (title ? `${id}_title` : void 0),
 		children: ({children, selected}) => {
 			if (!Array.isArray(children)) return [];
 
@@ -305,11 +292,11 @@ const DropdownBase = kind({
 		)
 	},
 
-	render: ({children, direction, disabled, onClose, onKeyDown, onOpen, onSelect, open, placeholder, selected, size, title, width, ...rest}) => {
+	render: ({ariaLabelledBy, children, direction, disabled, onClose, onOpen, onSelect, open, placeholder, selected, size, title, width, ...rest}) => {
 		delete rest.rtl;
 
 		const ariaProps = extractAriaProps(rest);
-		const popupProps = {'aria-live': null, children, onKeyDown, onSelect, selected, width, role: null};
+		const popupProps = {'aria-live': null, children, onSelect, selected, width, role: null};
 
 		// `ui/Group`/`ui/Repeater` will throw an error if empty so we disable the Dropdown and
 		// prevent Dropdown to open if there are no children.
@@ -317,11 +304,12 @@ const DropdownBase = kind({
 		const openDropdown = hasChildren && !disabled && open;
 
 		return (
-			<div role="region" {...rest}>
+			<div role="region" aria-labelledby={ariaLabelledBy} {...rest}>
 				{title}
 				<DropdownButton
 					direction={direction}
 					disabled={hasChildren ? disabled : true}
+					focusEffect="static"
 					icon={openDropdown ? 'arrowlargeup' : 'arrowlargedown'}
 					popupProps={popupProps}
 					popupComponent={DropdownList}
@@ -329,6 +317,7 @@ const DropdownBase = kind({
 					onClose={onClose}
 					open={openDropdown}
 					size={size}
+					spotlightRestrict="self-only"
 					{...ariaProps}
 				>
 					{placeholder}
@@ -354,9 +343,11 @@ const DropdownBase = kind({
  * @public
  */
 const DropdownDecorator = compose(
-	Pure({propComparators: {
-		children: compareChildren
-	}}),
+	Pure({
+		propComparators: {
+			children: compareChildren
+		}
+	}),
 	I18nContextDecorator({
 		rtlProp: 'rtl'
 	}),