+++ Input/NumberField.js
@@ -1,5 +1,5 @@
 import kind from '@enact/core/kind';
-import {handle, adaptEvent, forward} from '@enact/core/handle';
+import {handle, adaptEvent, forward, returnsTrue} from '@enact/core/handle';
 import PropTypes from 'prop-types';
 import React from 'react';
 import compose from 'ramda/src/compose';
@@ -30,6 +30,7 @@ const NumberCell = kind({
 	propTypes: /** @lends sandstone/Input.NumberCell.prototype */ {
 		active: PropTypes.bool,
 		children: PropTypes.string,
+		disabled: PropTypes.bool,
 		password: PropTypes.bool,
 		passwordIcon: PropTypes.string
 	},
@@ -66,7 +67,9 @@ const NumberFieldBase = kind({
 	name: 'NumberField',
 
 	propTypes: {
+		announce: PropTypes.func,
 		css: PropTypes.object,
+		disabled: PropTypes.bool,
 		invalid: PropTypes.bool,
 		invalidMessage: PropTypes.string,
 		maxLength: PropTypes.number,
@@ -96,14 +99,17 @@ const NumberFieldBase = kind({
 			adaptEvent(
 				({key}, {maxLength, value}) => ({value: normalizeValue(`${value}${key}`, maxLength)}),
 				handle(
+					returnsTrue(({value}, {announce, type}) => {
+						announce(type === 'password' ? $L('hidden') : String(value).substr(-1));
+					}),
 					// In case onAdd was run in the short period between the last onComplete and this invocation, just bail out
 					({value: updatedValue}, {maxLength, value}) => (normalizeValue(updatedValue, maxLength) !== normalizeValue(value, maxLength)),
 					forward('onChange'),
 					// Check the length of the new value and return true (pass/proceed) if it is at or above max-length
-					({value: updatedValue}, {maxLength, minLength}) => {
+					({value: updatedValue}, {maxLength, minLength, numberInputField}) => {
 						const
 							updatedLength = normalizeValue(updatedValue, maxLength).length,
-							autoSubmit = minLength === maxLength;
+							autoSubmit = getSeparated(numberInputField, maxLength) && minLength === maxLength;
 						return autoSubmit && updatedLength >= maxLength;
 					},
 					forward('onComplete')
@@ -111,6 +117,7 @@ const NumberFieldBase = kind({
 			)
 		),
 		onRemove: handle(
+			returnsTrue((ev, {announce}) => announce($L('Back Space'))),
 			adaptEvent(
 				(ev, {maxLength, value}) => ({value: normalizeValue(value, maxLength).toString().slice(0, -1)}),
 				forward('onChange')
@@ -140,11 +147,11 @@ const NumberFieldBase = kind({
 				);
 			}
 		},
-		submitButton: ({css, invalid, maxLength, minLength, onSubmit, value}) => {
-			const disabled = invalid || (normalizeValue(value, maxLength).toString().length < minLength);
+		submitButton: ({css, disabled, invalid, maxLength, minLength, onSubmit, value, numberInputField}) => {
+			const isDisabled = disabled || invalid || (normalizeValue(value, maxLength).toString().length < minLength);
 
-			if (minLength !== maxLength) {
-				return <Button className={css.submitButton} disabled={disabled} onClick={onSubmit}>{$L('Submit')}</Button>;
+			if (minLength !== maxLength || !getSeparated(numberInputField, maxLength)) {
+				return <Button className={css.submitButton} disabled={isDisabled} onClick={onSubmit}>{$L('Submit')}</Button>;
 			} else {
 				return null;
 			}
@@ -157,8 +164,9 @@ const NumberFieldBase = kind({
 		}
 	},
 
-	render: ({css, invalidTooltip, maxLength, numberInputField, onAdd, onRemove, showKeypad, submitButton, type, value, ...rest}) => {
+	render: ({css, disabled, invalidTooltip, maxLength, numberInputField, onAdd, onRemove, showKeypad, submitButton, type, value, ...rest}) => {
 		const password = (type === 'password');
+		delete rest.announce;
 		delete rest.invalid;
 		delete rest.invalidMessage;
 		delete rest.minLength;
@@ -174,8 +182,6 @@ const NumberFieldBase = kind({
 			const items = new Array(maxLength).fill('');
 			field = (
 				<Repeater
-					aria-label={!password ? values.join(' ') : null}
-					aria-live="polite"
 					{...rest}
 					component={Layout}
 					childComponent={Cell}
@@ -185,6 +191,7 @@ const NumberFieldBase = kind({
 						active: index <= value.length,
 						children: values[index],
 						component: NumberCell,
+						disabled,
 						key: `key-${index}`,
 						password,
 						shrink: true
@@ -193,7 +200,7 @@ const NumberFieldBase = kind({
 			);
 		} else {
 			field = (
-				<div {...rest}>
+				<div {...rest} disabled={disabled}>
 					{password ? convertToPasswordFormat(value) : value}
 				</div>
 			);
@@ -206,7 +213,7 @@ const NumberFieldBase = kind({
 					{invalidTooltip}
 				</div>
 				<br />
-				{showKeypad ? <Keypad onAdd={onAdd} onRemove={onRemove} /> : null}
+				{showKeypad ? <Keypad aria-label=" " disabled={disabled} onAdd={onAdd} onRemove={onRemove} /> : null}
 				{submitButton}
 			</React.Fragment>
 		);