+++ samples/qa-a11y/src/views/A11yDecorator.js
@@ -6,15 +6,15 @@ import React from 'react';
 const A11yButton = A11yDecorator(Button);
 
 const A11yDecoratorView = () => (
-	<div>
+	<>
 		<Heading showLine>Button Examples</Heading>
 		<p><strong>accessibilityPreHint=Button Examples</strong></p>
-		<A11yButton size="small" accessibilityPreHint="Button Examples">Easy</A11yButton>
+		<A11yButton accessibilityPreHint="Button Examples">Easy</A11yButton>
 		<p><strong>accessibilityHint=Hint</strong></p>
-		<A11yButton size="small" accessibilityHint="Hint" backgroundOpacity="translucent">Medium</A11yButton>
+		<A11yButton accessibilityHint="Hint" backgroundOpacity="transparent">Medium</A11yButton>
 		<p><strong>aria-label=accessibility</strong></p>
-		<A11yButton size="small" aria-label="accessibility" backgroundOpacity="transparent">Hard</A11yButton>
-	</div>
+		<A11yButton aria-label="accessibility" backgroundOpacity="transparent">Hard</A11yButton>
+	</>
 );
 
 export default A11yDecoratorView;