+++ samples/qa-a11y/src/views/Dropdown.js
@@ -1,7 +1,5 @@
 import Dropdown from '@enact/sandstone/Dropdown';
-import Layout, {Cell} from '@enact/ui/Layout';
 import React from 'react';
-import Scroller from '@enact/sandstone/Scroller';
 
 const list = [
 	{children: 'Option1', key: 'item1', 'aria-label': 'aria label 1'},
@@ -36,18 +34,21 @@ class A11yDropdown extends React.Component {
 }
 
 const DropdownView = () => (
-	<Layout orientation="vertical">
-		<Cell component={Scroller} focusableScrollbar>
-			<Dropdown placeholder="Dropdown" title="String Array children">
-				{['Option1', 'Option2', 'Option3']}
-			</Dropdown>
-			<br />
-			<br />
-			<A11yDropdown placeholder="Dropdown" title="Object Array children with aria-label">
-				{list}
-			</A11yDropdown>
-		</Cell>
-	</Layout>
+	<>
+		<Dropdown placeholder="Dropdown without title">
+			{['Option1', 'Option2', 'Option3']}
+		</Dropdown>
+		<br />
+		<br />
+		<Dropdown placeholder="Dropdown" title="String Array children">
+			{['Option1', 'Option2', 'Option3']}
+		</Dropdown>
+		<br />
+		<br />
+		<A11yDropdown placeholder="Dropdown" title="Object Array children with aria-label">
+			{list}
+		</A11yDropdown>
+	</>
 );
 
 export default DropdownView;